refactor: security 적용 버전으로 리팩토링
