+++ src/main/java/org/example/broong/config/WebConfig.java
@@ -1,7 +1,7 @@
 package org.example.broong.config;
 
 import lombok.RequiredArgsConstructor;
-import org.example.broong.interceptor.OwnerAccessInterceptor;
+import org.example.broong.security.interceptor.OwnerAccessInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
