+++ src/main/java/org/example/broong/domain/auth/service/AuthService.java
@@ -9,7 +9,7 @@ import java.util.Date;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.broong.security.jwt.JwtService;
-import org.example.broong.security.RedisDao;
+import org.example.broong.security.auth.RedisDao;
 import org.example.broong.domain.auth.dto.request.AuthRequestDto;
 import org.example.broong.domain.user.enums.LoginType;
 import org.example.broong.global.exception.ApiException;
@@ -62,11 +62,19 @@ public class AuthService {
                 () -> new ApiException(HttpStatus.BAD_REQUEST, NO_RESOURCE, "액세스 토큰이 존재하지 않습니다.")
         );
 
-        if (!jwtService.isValidToken(accessToken, "access")) {
-            throw new ApiException(HttpStatus.UNAUTHORIZED, NO_RESOURCE, "유효하지 않은 액세스 토큰입니다.");
-        }
+//        if (!jwtService.isValidToken(accessToken, "access")) {
+//            throw new ApiException(HttpStatus.UNAUTHORIZED, NO_RESOURCE, "유효하지 않은 액세스 토큰입니다.");
+//        }
 
         Claims claims = jwtService.extractClaims(accessToken, "access");
+
+        String email = jwtService.extractEmail(claims);
+
+        if (redisDao.hasKey(email)) {
+            redisDao.deleteRefreshToken(email);
+        } else {
+            throw new ApiException(HttpStatus.FORBIDDEN, INVALID_PARAMETER, "이미 로그아웃한 유저입니다.");
+        }
 
         // 현재 만료시간이 얼마큼 남았는지 확인
         Date expiration = claims.getExpiration();
@@ -82,13 +90,7 @@ public class AuthService {
 
         log.info("블랙리스트 {}",redisDao.getBlackList(accessToken));
 
-        String email = jwtService.extractEmail(claims);
 
-        if (redisDao.hasKey(email)) {
-            redisDao.deleteRefreshToken(email);
-        } else {
-            throw new ApiException(HttpStatus.FORBIDDEN, INVALID_PARAMETER, "이미 로그아웃한 유저입니다.");
-        }
 
         log.info("리프레시 삭제 여부 {}",redisDao.hasKey(email));
